RAVE 0.602e
- Mes retocs al sistema del fade in/out
- Creada nova finestra d'opcions per anar movent alla totes les opcions de forma mes ordenada

+++ Reproductor de Audio y Video Extendido/VentanaPrincipal.h
@@ -277,7 +277,7 @@ class VentanaPrincipal : public DWL::DVentana {
 								// Variable que determina si se ha entrado en modo patalla completa
 	BOOL                       _PantallaCompleta;		
 								// Ultimo tick del sistema al presionar el boton del mouse
-	DWORD                      _BotonExMouseDownTick;	
+	ULONGLONG                  _BotonExMouseDownTick;	
 								// Total de medios en la fase 2 del analisis
 	UINT                       _MaximoTotalMedios2;	
 								// Valor actual de los medios analizados en la fase 2
